pyproject.toml
[project]
name = "vinyller"
version = "1.0.0"
description = "Vinyller is a music player with library, metatags and encyclopedia management"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Maxim Moshkin", email = "hellomaxcreations@gmail.com"}
]
license = {text = "GPL-3.0-or-later"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPL-3.0-or-later)",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "PyQt6>=6.10.2",
    "mutagen>=1.47.0",
    "requests>=2.33.0",
    "urllib3>=2.6.3",
    "Pillow>=12.1.1",
    "send2trash>=2.1.0",
    "pyobjc-framework-Cocoa>=12.1; sys_platform == 'darwin'",
    "pyobjc-framework-MediaPlayer>=12.1; sys_platform == 'darwin'",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.18.0",
]

[project.scripts]
vinyller = "vinyller:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.wheel.force-include]
"assets" = "assets"
"translations" = "translations"
"LICENSE" = "LICENSE"
"vinyller.py" = "vinyller.py"

[project.urls]
"Homepage" = "https://github.com/maxcreations/vinyller"
"Bug Tracker" = "https://github.com/maxcreations/vinyller/issues"